refactor: extract shared test helpers into CustomerCommandTestHelper trait
